+++ src/generator.rs
@@ -1,0 +1,6 @@
+// Copyright (C) 2025 FZI Forschungszentrum Informatik
+// SPDX-License-Identifier: Apache-2.0
+//! Payload generators
+
+pub mod error;
+pub mod state;